Hospital-Microservices/patient-service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic
httpx
python-jose
python-multipart
